Finished separating repeat detection elimination into two scripts with a shared core.
