+++ .gitignore
@@ -26,3 +26,4 @@ demo/CameraTrapAssets
 .webassets-cache/
 demo/static/uploads
 demo/static/results
+.spyproject/
